WIP removing routing context
