+++ src/components/presentation/Progress.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { getLabelId } from 'src/components/label/Label';
 import classes from 'src/components/presentation/Progress.module.css';
 import { useLanguage } from 'src/features/language/useLanguage';
-import { useNavigationParam } from 'src/features/routing/AppRoutingContext';
+import { useNavigationParam } from 'src/hooks/navigation';
 import { useNavigatePage } from 'src/hooks/useNavigatePage';
 
 export const Progress = () => {
